LayoutTests/fast/js/dfg-arith-add-overflow-check-elimination-tower-of-large-numbers-expected.txt

Tests that if we have a tower of large numerical constants being added to each other, the DFG knows that a sufficiently large tower may produce a large enough value that overflow check elimination must be careful.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(1, 2) is 0
PASS bar(2147483645, 2147483644) is -10
PASS successfullyParsed is true

TEST COMPLETE
